requirements.txt
pandas==1.1.5
beautifulsoup4==4.9.3
lxml==4.6.3
numpy==1.19.5
DateTime==4.3
requests==2.23.0
requests-oauthlib==1.3.0
matplotlib==3.2.2
six==1.15.0
Pillow==7.1.2
openpyxl==2.5.9
python-decouple==3.6